requirements.txt
fastapi
uvicorn[standard]
httpx
jinja2
pytest
pytest-mock
python-dotenv